implemented error exception handling in game controller and game service

+++ backend/app/src/game/game.service.ts
@@ -1,19 +1,26 @@
-import { Injectable, NotFoundException } from "@nestjs/common";
+import {
+  BadRequestException,
+  Injectable,
+  Logger,
+  NotFoundException,
+} from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
 import { Game } from "./entities/game.entity";
 import { User } from "src/user/user.entity";
+import { UserService } from "src/user/user.service";
 import { CreateGameDto } from "./dto/create-game.dto";
 import { UpdateGameDto } from "./dto/update-game.dto";
 
 @Injectable()
 export class GameService {
+  private readonly logger = new Logger(GameService.name);
+
   constructor(
     @InjectRepository(Game)
     private readonly gameRepository: Repository<Game>,
 
-    @InjectRepository(User)
-    private readonly userRepository: Repository<User>,
+    private readonly userService: UserService,
   ) {}
 
   async findAll() {
@@ -30,13 +37,33 @@ export class GameService {
     return game;
   }
 
+  /* need to also add check to ensure they are not in playing status */
   async create(createGameDto: CreateGameDto) {
-    /* validate that both users exist before creating the game */
+    const playerOne = await this.userService.findUserById(
+      createGameDto.playerOne,
+    );
+    if (playerOne === null) {
+      this.logger.debug("playerOne does not exist in database");
+      throw new NotFoundException();
+    }
+    const playerTwo = await this.userService.findUserById(
+      createGameDto.playerTwo,
+    );
+    if (playerTwo === null) {
+      this.logger.debug("playerTwo does not exist in database");
+      throw new NotFoundException();
+    }
     return await this.gameRepository.save(createGameDto);
   }
 
   async update(updateGameDto: UpdateGameDto) {
     if (updateGameDto.id !== undefined) {
+      if ((await this.findOne(updateGameDto.id)) == null) {
+        this.logger.debug("game does not exist, unable to update");
+        throw new NotFoundException(
+          "Unable to update game because game does not exist",
+        );
+      }
       await this.gameRepository.update(updateGameDto.id, updateGameDto);
 
       await this.gameRepository
@@ -54,11 +81,15 @@ export class GameService {
       return this.findOne(updateGameDto.id);
     }
     throw new NotFoundException(
-      "Unable to update game because it does not exist",
+      "Unable to update game because no data received from dto",
     );
   }
 
   async remove(gameId: number) {
+    if (this.findOne(gameId) == null) {
+      this.logger.debug("game does not exist, unable to delete");
+      throw new NotFoundException("game does not exist, unable to delete");
+    }
     await this.gameRepository.delete(gameId);
   }
 }
